Extract sass code out of site/lib.rs
