+++ components/site/src/lib.rs
@@ -1,13 +1,13 @@
+pub mod sass;
 pub mod sitemap;
 
 use std::collections::HashMap;
-use std::fs::{copy, create_dir_all, remove_dir_all};
+use std::fs::{copy, remove_dir_all};
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex, RwLock};
 
 use glob::glob;
 use rayon::prelude::*;
-use sass_rs::{compile_file, Options as SassOptions, OutputStyle};
 use serde_derive::Serialize;
 use tera::{Context, Tera};
 
@@ -47,15 +47,19 @@ pub struct Site {
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize)]
-struct SerializedTaxonomyItem<'a> {
+struct SerializedFeedTaxonomyItem<'a> {
     name: &'a str,
     slug: &'a str,
     permalink: &'a str,
 }
 
-impl<'a> SerializedTaxonomyItem<'a> {
+impl<'a> SerializedFeedTaxonomyItem<'a> {
     pub fn from_item(item: &'a TaxonomyItem) -> Self {
-        SerializedTaxonomyItem { name: &item.name, slug: &item.slug, permalink: &item.permalink }
+        SerializedFeedTaxonomyItem {
+            name: &item.name,
+            slug: &item.slug,
+            permalink: &item.permalink,
+        }
     }
 }
 
@@ -748,12 +752,12 @@ impl Site {
         if let Some(ref theme) = self.config.theme {
             let theme_path = self.base_path.join("themes").join(theme);
             if theme_path.join("sass").exists() {
-                self.compile_sass(&theme_path)?;
+                sass::compile_sass(&theme_path, &self.output_path)?;
             }
         }
 
         if self.config.compile_sass {
-            self.compile_sass(&self.base_path)?;
+            sass::compile_sass(&self.base_path, &self.output_path)?;
         }
 
         if self.config.build_search_index {
@@ -838,71 +842,6 @@ impl Site {
         create_file(&self.output_path.join("elasticlunr.min.js"), search::ELASTICLUNR_JS)?;
 
         Ok(())
-    }
-
-    pub fn compile_sass(&self, base_path: &Path) -> Result<()> {
-        ensure_directory_exists(&self.output_path)?;
-
-        let sass_path = {
-            let mut sass_path = PathBuf::from(base_path);
-            sass_path.push("sass");
-            sass_path
-        };
-
-        let mut options = SassOptions::default();
-        options.output_style = OutputStyle::Compressed;
-        let mut compiled_paths = self.compile_sass_glob(&sass_path, "scss", &options.clone())?;
-
-        options.indented_syntax = true;
-        compiled_paths.extend(self.compile_sass_glob(&sass_path, "sass", &options)?);
-
-        compiled_paths.sort();
-        for window in compiled_paths.windows(2) {
-            if window[0].1 == window[1].1 {
-                bail!(
-                    "SASS path conflict: \"{}\" and \"{}\" both compile to \"{}\"",
-                    window[0].0.display(),
-                    window[1].0.display(),
-                    window[0].1.display(),
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    fn compile_sass_glob(
-        &self,
-        sass_path: &Path,
-        extension: &str,
-        options: &SassOptions,
-    ) -> Result<Vec<(PathBuf, PathBuf)>> {
-        let glob_string = format!("{}/**/*.{}", sass_path.display(), extension);
-        let files = glob(&glob_string)
-            .expect("Invalid glob for sass")
-            .filter_map(|e| e.ok())
-            .filter(|entry| {
-                !entry.as_path().components().any(|c| c.as_os_str().to_string_lossy().starts_with('_'))
-            })
-            .collect::<Vec<_>>();
-
-        let mut compiled_paths = Vec::new();
-        for file in files {
-            let css = compile_file(&file, options.clone())?;
-
-            let path_inside_sass = file.strip_prefix(&sass_path).unwrap();
-            let parent_inside_sass = path_inside_sass.parent();
-            let css_output_path = self.output_path.join(path_inside_sass).with_extension("css");
-
-            if parent_inside_sass.is_some() {
-                create_dir_all(&css_output_path.parent().unwrap())?;
-            }
-
-            create_file(&css_output_path, &css)?;
-            compiled_paths.push((path_inside_sass.to_owned(), css_output_path));
-        }
-
-        Ok(compiled_paths)
     }
 
     fn render_alias(&self, alias: &str, permalink: &str) -> Result<()> {
@@ -1138,7 +1077,7 @@ impl Site {
 
         if let Some((taxonomy, item)) = taxonomy_and_item {
             context.insert("taxonomy", taxonomy);
-            context.insert("term", &SerializedTaxonomyItem::from_item(item));
+            context.insert("term", &SerializedFeedTaxonomyItem::from_item(item));
         }
 
         let feed = &render_template(feed_filename, &self.tera, context, &self.config.theme)?;
@@ -1292,3 +1231,5 @@ impl Site {
             .collect::<Result<()>>()
     }
 }
+
+impl Site {}
